requirements.txt
# Core
Django==5.0.9
python-dotenv==1.0.1
psycopg2-binary==2.9.10

# Telegram Bot (aiogram вместо python-telegram-bot)
aiogram==3.13.1
aiohttp==3.10.5
magic-filter==1.0.12

# Redis & Celery
redis==5.1.1
celery==5.4.0
django-celery-beat==2.6.0
django-celery-results==2.5.1

# Database
django-redis==5.4.0

# API & Serialization  
djangorestframework==3.15.2
django-cors-headers==4.4.0

# AI/NLP для обработки текста
openai==1.55.3
google-generativeai==0.7.2
langdetect==1.0.9
pymorphy3==2.0.2
pymorphy3-dicts-ru==2.4.417150.4580142

# PDF Generation
reportlab==4.2.2
matplotlib==3.9.2
Pillow==11.0.0
playwright==1.41.0
jinja2==3.1.4

# Utils
python-dateutil==2.9.0
pytz==2024.1
babel==2.16.0
aiofiles==24.1.0

# Excel/Data Processing
openpyxl==3.1.5
pandas==2.2.2
numpy==2.1.3

# Development
ipython==8.26.0
django-extensions==3.2.3

# Production
gunicorn==22.0.0
whitenoise==6.8.2
dj-database-url==2.2.0

# Voice Recognition
SpeechRecognition==3.10.4
pydub==0.25.1

# Testing
pytest==8.3.2
pytest-django==4.8.0
pytest-asyncio==0.23.8
factory-boy==3.3.0